+++ tests/container-scripts/integration-tests.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+NODES=$1
+TEST_TYPE=$2
 set -eu
 
 FILE=/contracts
@@ -14,4 +16,4 @@ killall -9 test-runner
 set -e
 
 pushd /peggy/orchestrator/test_runner
-RUST_BACKTRACE=full NO_GAS_OPT=1 RUST_LOG=INFO PATH=$PATH:$HOME/.cargo/bin cargo run --release --bin test-runner
+RUST_BACKTRACE=full TEST_TYPE=$TEST_TYPE NO_GAS_OPT=1 RUST_LOG=INFO PATH=$PATH:$HOME/.cargo/bin cargo run --release --bin test-runner
