+++ tests/container-scripts/reload-code.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
-set -eux
-
 # Number of validators to start
 NODES=$1
+# what test to execute
+TEST_TYPE=$2
+set -eux
 
 # Stop any currently running peggy and eth processes
 pkill peggyd || true # allowed to fail
@@ -24,7 +25,7 @@ tests/container-scripts/run-testnet.sh $NODES
 
 # deploy the ethereum contracts
 pushd /peggy/orchestrator/test_runner
-DEPLOY_CONTRACTS=1 RUST_BACKTRACE=full NO_GAS_OPT=1 RUST_LOG=INFO PATH=$PATH:$HOME/.cargo/bin cargo run --release --bin test-runner
+DEPLOY_CONTRACTS=1 RUST_BACKTRACE=full TEST_TYPE=$TEST_TYPE NO_GAS_OPT=1 RUST_LOG=INFO PATH=$PATH:$HOME/.cargo/bin cargo run --release --bin test-runner
 
 # This keeps the script open to prevent Docker from stopping the container
 # immediately if the nodes are killed by a different process
